requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
elasticsearch==8.12.0
strawberry-graphql==0.219.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0